Refactored, fixed bugs in BCItems
