+++ src/main/java/net/buildcraftr/buildcraft_revolution/items/Chipset/ChipsetQuartz.java
@@ -1,4 +1,4 @@
-package net.buildcraftr.buildcraft_revolution.items;
+package net.buildcraftr.buildcraft_revolution.items.Chipset;
 
 import net.minecraft.item.Item;
