+++ src/main/java/net/buildcraftr/buildcraft_revolution/registry/BCItems.java
@@ -2,6 +2,8 @@ package net.buildcraftr.buildcraft_revolution.registry;
 
 import net.buildcraftr.buildcraft_revolution.BuildcraftRevolution;
 import net.buildcraftr.buildcraft_revolution.items.*;
+import net.buildcraftr.buildcraft_revolution.items.Chipset.*;
+import net.buildcraftr.buildcraft_revolution.items.Gear.*;
 import net.buildcraftr.buildcraft_revolution.items.Paintbrush.*;
 import net.buildcraftr.buildcraft_revolution.tab.CreativeTabs;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
@@ -12,12 +14,16 @@ import net.minecraft.util.registry.Registry;
 public class BCItems {
 
     // Register Items -------------------------------------------------------------
+    //Misc
     public static final Item BC_WRENCH = registerItem("wrench", new Wrench(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS).maxCount(1)));
+    public static final Item BC_PIPE_SEALANT = registerItem("pipe_sealant", new PipeSealant(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
+    //Gear
     public static final Item BC_GEAR_WOOD = registerItem("gear_wood", new GearWood(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
-    public static final Item BC_GEAR_STONE = registerItem("gear_stone", new GearWood(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
-    public static final Item BC_GEAR_IRON = registerItem("gear_iron", new GearWood(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
-    public static final Item BC_GEAR_GOLD = registerItem("gear_gold", new GearWood(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
-    public static final Item BC_GEAR_DIAMOND = registerItem("gear_diamond", new GearWood(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
+    public static final Item BC_GEAR_STONE = registerItem("gear_stone", new GearStone(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
+    public static final Item BC_GEAR_IRON = registerItem("gear_iron", new GearIron(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
+    public static final Item BC_GEAR_GOLD = registerItem("gear_gold", new GearGold(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
+    public static final Item BC_GEAR_DIAMOND = registerItem("gear_diamond", new GearDiamond(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
+    //Paintbrush
     public static final Item BC_PAINTBRUSH_CLEAN = registerItem("paintbrush_clean", new PaintbrushClean(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS).maxCount(1)));
     public static final Item BC_PAINTBRUSH_WHITE = registerItem("paintbrush_white", new PaintbrushWhite(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS).maxCount(1)));
     public static final Item BC_PAINTBRUSH_ORANGE = registerItem("paintbrush_orange", new PaintbrushOrange(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS).maxCount(1)));
@@ -35,7 +41,7 @@ public class BCItems {
     public static final Item BC_PAINTBRUSH_GREEN = registerItem("paintbrush_green", new PaintbrushGreen(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS).maxCount(1)));
     public static final Item BC_PAINTBRUSH_RED = registerItem("paintbrush_red", new PaintbrushRed(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS).maxCount(1)));
     public static final Item BC_PAINTBRUSH_BLACK = registerItem("paintbrush_black", new PaintbrushBlack(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS).maxCount(1)));
-    public static final Item BC_PIPE_SEALANT = registerItem("pipe_sealant", new GearWood(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
+    //Chipset
     public static final Item BC_CHIPSET_REDSTONE = registerItem("chipset_redstone", new ChipsetRedstone(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
     public static final Item BC_CHIPSET_IRON = registerItem("chipset_iron", new ChipsetIron(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
     public static final Item BC_CHIPSET_GOLD = registerItem("chipset_gold", new ChipsetGold(new FabricItemSettings().group(CreativeTabs.TAB_ITEMS)));
